+++ src/client/MarketClient.ts
@@ -1,9 +1,10 @@
 import { AvanzaClient } from "./AvanzaClient";
-import { SearchResponse } from "./market_types";
+import { InstrumentResponse, SearchResponse } from "./market_types";
 import { InstrumentType } from "./types";
 
 const MARKET_PATHS = {
   SEARCH: "/_cqbe/search/global-search/global-search-template",
+  GET_INSTRUMENT: "/_api/market-guide/{instrument}/{id}",
 };
 
 export class MarketClient {
@@ -21,6 +22,25 @@ export class MarketClient {
       throw new Error(JSON.stringify(errorMessage));
     }
     const data = (await response.json()) as SearchResponse;
+    return data;
+  }
+
+  async getInstrument(
+    instrumentType: InstrumentType,
+    instrumentId: string
+  ): Promise<InstrumentResponse> {
+    const response = await this.client.get(
+      MARKET_PATHS.GET_INSTRUMENT.replace(
+        "{instrument}",
+        instrumentType.toLowerCase()
+      ).replace("{id}", instrumentId),
+      {}
+    );
+    if (!response.ok) {
+      const errorMessage = await response.text();
+      throw new Error(JSON.stringify(errorMessage));
+    }
+    const data = (await response.json()) as InstrumentResponse;
     return data;
   }
 }